+++ examples/multiple_base_routes/lib/main.dart
@@ -6,6 +6,7 @@ import 'package:multiple_base_routes/data.dart';
 import 'package:multiple_base_routes/deep_links.dart';
 import 'package:multiple_base_routes/model.dart';
 import 'package:multiple_base_routes/widgets/artist_page.dart';
+import 'package:multiple_base_routes/widgets/authentication_page.dart';
 import 'package:multiple_base_routes/widgets/error_page.dart';
 import 'package:multiple_base_routes/widgets/favorites_page.dart';
 import 'package:multiple_base_routes/widgets/library_page.dart';
@@ -13,8 +14,6 @@ import 'package:multiple_base_routes/widgets/login_page.dart';
 import 'package:multiple_base_routes/widgets/song_page.dart';
 import 'package:multiple_base_routes/widgets/splash_page.dart';
 import 'package:multiple_base_routes/widgets/user_page.dart';
-
-final navigatorKey = GlobalKey<NavigatorState>();
 
 void main() => runApp(
   // State management of your choice
@@ -37,29 +36,30 @@ class MusicApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) => DeepLinkMaterialApp(
     // This is where the magic happens
-    navigationBuilder: (baseDispatcher) => baseDispatcher
+    navigation: (context) => Dispatcher()
       // RouteNotFound error page
       ..exception<RouteNotFound>((context, exception, path) => [ErrorDL<RouteNotFound>(exception)])
-      ..value<RouteNotFound, ErrorDL<RouteNotFound>>((context, exception, path) => ErrorPage(exception))
+      ..value<RouteNotFound, ErrorDL<RouteNotFound>>((exception, path) => ErrorPage(exception))
       // Unauthenticated login page
       ..exception<Unauthenticated>((context, exception, path) => [LoginDL()])
-      ..path<LoginDL>((context, path) => LoginPage())
-
+      ..path<LoginDL>((path) => LoginPage())
+      // The rest of the app
       ..path<LibraryDL>(
-        (context, path) => LibraryPage(),
-        subNavigationBuilder: (dispatcher) => dispatcher
+        (path) => LibraryPage(),
+        subNavigation: (context) => Dispatcher()
           ..value<Artist, ArtistDL>(
-            (context, artist, path) => ArtistPage(artist: artist),
-            subNavigationBuilder: (dispatcher) => dispatcher
-              ..value<Song, SongDL>((context, song, path) => SongPage(song: song)), // TODO: static extension method
+            (artist, path) => ArtistPage(artist: artist),
+            subNavigation: (context, artist) => Dispatcher()..value<Song, SongDL>((song, path) => SongPage(song: song)), // TODO: static extension method
           ),
       )
       ..path<FavoritesDL>(
-        (context, path) => FavoritesPage(),
-        subNavigationBuilder: (dispatcher) => dispatcher
-          ..value<Song, SongDL>((context, song, path) => SongPage(song: song)), // TODO: static extension method
+        (path) => FavoritesPage(),
+        subNavigation: (context) => Dispatcher()..value<Song, SongDL>((song, path) => SongPage(song: song)), // TODO: static extension method
       )
-      ..path<UserDL>((context, path) => UserPage()),
+      ..path<UserDL>(
+        (path) => UserPage(),
+        subNavigation: (context) => Dispatcher()..path<AuthenticationDL>((path) => AuthenticationPage()),
+      ),
     defaultRoute: [LibraryDL()],
     splashScreen: SplashPage(),
     // Optionally specify always visible widgets