+++ examples/multiple_base_routes/test_driver/app_test.dart
@@ -13,7 +13,7 @@ Future<void> main() {
       ProgressReporter(),
       TestRunSummaryReporter(),
     ]
-    ..stepDefinitions = [TextDisappears(), TitleIs()]
+    ..stepDefinitions = [GoBack(), TextDisappears(), TitleIs()]
     ..restartAppBetweenScenarios = true
     ..targetAppPath = "test_driver/app.dart"
     ..exitAfterTestRun = true;