+++ examples/multiple_base_routes/test_driver/steps/steps.dart
@@ -1,2 +1,3 @@
+export 'go_back.dart';
 export 'text_disappears.dart';
 export 'title_is.dart';