+++ lib/src/pop_observer.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:deep_link_navigation/deep_link_navigation.dart';
 
-/// Notifies [DeepLinkNavigator] of a pop that occurred from the native flutter navigator.
+/// Notifies [DeepLinkNavigator] of a pop that occurred from the native Flutter navigator.
 /// Handles android back button, back arrow, and [DeepLinkNavigator.pop()] calls native pop.
 ///
 /// eg. Navigator(observers: [PopObserver()])