+++ test/deep_link_navigation_test.dart
@@ -1,2 +1,2 @@
-// TODO: deep link navigator unit tests (mock navigator)
+// TODO: deep link navigator unit tests (mock native navigator)
 // navigating to same route